+++ CHANGELOG.md
@@ -2,12 +2,18 @@
 
 ## 5.0.0 (IN PROGRESS)
 
+### Breaking changes
+
+- Dashboard variables in Show If, Disabled and Options replace automatically. Function `replaceVariables` is not required and removed.
+
 ### Features / Enhancements
 
 - Updated change value behavior for custom option (#562)
 - Added form-element of type color (#561)
 - Updated new lines in Text Area and Code Editor (#563)
 - Added support variables in data sources (#564)
+- Updated packages for Code Editor (#569)
+- Updated element helpers (#570)
 
 ## 4.9.0 (2024-11-16)